feat: allow callers of launch to react to errors
Adds a functional interface as a parameter to launch. This SAM is invoked with
the result of the launch call. If the launch fails, the result object will
contain information as to why it failed. If the result is successful there is
no extra data.

The default SAM will just log errors to logcat.